requirements.txt
# app
aiofiles
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.0
azure-storage-blob==12.20.0
beartype==0.18.5
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.8
cycler==0.12.1
dask==2024.7.0
dask-expr==1.1.7
datashaper==0.0.49
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
environs==11.0.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastparquet==2024.5.0
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gensim==4.3.2
gradio==4.38.1
gradio_client==1.1.0
#graphrag==0.1.1
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
hyppo==0.4.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
lancedb==0.9.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
ollama==0.2.1
openai==1.35.14
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
partd==1.4.2
patsy==0.5.6
pillow==10.4.0
plotly==5.22.0
portalocker==2.10.1
POT==0.9.4
psutil==6.0.0
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.13.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.2
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
swifter==1.4.0
tenacity==8.5.0
textual==0.70.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.1
#uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2


# api
fastapi
uvicorn
pandas
tavily-python
pydantic
python-dotenv
asyncio
aiohttp
scikit-learn
seaborn
spacy
# transformers
# API情况下需要单独安装
# graphrag==0.1.1

# cu121
# torch
# torchvision
# torchaudio
